stillimage/requirements.txt
pillow==7.1.2
requests==2.24.0
